requirements.txt
asgiref==3.8.1
Django==5.1.4
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
django-filter==25.1
django-jazzmin==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2024.2
